+++ test/com/ansorgit/plugins/bash/lang/lexer/BashLexerTest.java
@@ -996,6 +996,11 @@ public class BashLexerTest {
     }
 
     @Test
+    public void testIssue272() throws Exception {
+        testTokenization("${#array_var[@]}", DOLLAR, LEFT_CURLY, PARAM_EXPANSION_OP_HASH, WORD, LEFT_SQUARE, PARAM_EXPANSION_OP_AT, RIGHT_SQUARE, RIGHT_CURLY);
+    }
+
+    @Test
     public void testTrapLexing() {
         testTokenization("trap", TRAP_KEYWORD);
         testTokenization("trap -l", TRAP_KEYWORD, WHITESPACE, WORD);